backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.20
pandas==2.2.3
openpyxl==3.1.5
google-generativeai==0.8.3
aiosmtplib==3.0.1
slowapi==0.1.9
pydantic[email]==2.10.3
python-dotenv==1.0.1
email-validator==2.2.0
